Move dockerfiles to root of this project
DockerHub lets you build your images automatically from your git repo,
but context that is send to docker builder is not root of the repo but instead
it's wherever your Dockerfile is placed.

Basically instead of building like this:
docker build -f alpine/5.0/Dockerfile -t nooulaif/rails:latest .

They are doing something like this:
cd alpine/5.0/ && docker build -t nooulaif/rails:latest .

Because of that files that are shared across Dockerfiles and placed in root,
are not visible in builder and it throws an error.

The simple solution to that is to place all Dockerfiles in root of a project.