Change base images to official ruby images
Although ~150MB rails images are fun to build and play with,
they are not very practical.

Almost all more complex apps will require some sort of gems that needs to be build
and previous versions of this image couldn't affort this if it wanted to be as
thin as possible.

There are many positive sides with switching to official docker ruby images:
1) They are well maintained and updated regularly.
2) They provide debian based images which are belived to be more stable and secure.
3) They have many commonly used tools like ImageMagick preinstalled. (in buildpack-deps)
4) They open up the possibility to use DockerHub utility to automatically rebuild our images every time ruby gets an update.

The only downside is of course the size.